pyproject.toml
[tool.poetry]
name = "documentary-watch"
version = "0.1.2"
license = "GPL-3.0-or-later"
description = ""

authors = ["DelphineLemire <delphine.lemire@gmail.com>"]
readme = "docs/README.md"
packages = [{include = "documentary_watch"}]

[tool.poetry.dependencies]
python = "^3.9"

# Django
# ------------------------------------------------------------------------------
django = "4.1.7"  # https://www.djangoproject.com/
django-environ = "0.10.0" # https://github.com/joke2k/django-environ
#django-allauth ="0.52.0"  # https://github.com/pennersr/django-allauth

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "3.8.1"  # https://github.com/jazzband/django-debug-toolbar
django-extensions = "3.2.1" # https://github.com/django-extensions/django-extensions

# Code quality
# ------------------------------------------------------------------------------

mypy = "1.1.1"  # https://github.com/python/mypy
pre-commit = "3.1.1"  # https://github.com/pre-commit/pre-commit
ruff = "^0.0.254" # https://github.com/charliermarsh/ruff
deptry = "^0.8.0" # https://github.com/fpgmaas/deptry


[tool.poetry.group.test]  # This part can be left out

[tool.poetry.group.test.dependencies]
# Testing
# ------------------------------------------------------------------------------

pytest = "7.2.2"  # https://github.com/pytest-dev/pytest
pytest-cov = "4.0.0" # https://github.com/pytest-dev/pytest-cov
pytest-sugar = "0.9.6" # https://github.com/Teemu/pytest-sugar/

# Testing Djangp
pytest-django = "4.5.2" # https://github.com/pytest-dev/pytest-django
factory-boy = "3.2.1"  # https://github.com/FactoryBoy/factory_boy
django-coverage-plugin = "3.0.0" # https://github.com/nedbat/django_coverage_plugin


[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2" # https://github.com/mkdocs/mkdocs
mkdocs-material = "9.1.2" # https://github.com/squidfunk/mkdocs-material
pymdown-extensions = "9.10" # https://github.com/facelessuser/pymdown-extensions
mkdocstrings = {extras = ["python"], version = "^0.20.0"} # https://github.com/mkdocstrings/mkdocstrings


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Tools's parameters

# ------------------------------------------------------------------------------

[tool.mypy]
python_version = 3.9
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = "*/migrations.*"
# Django migrations should not produce any errors:
ignore_errors = true

[tool.ruff]
target-version = "py39"
line-length = 120
fix = false
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # LineTooLong
    "E501",
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.per-file-ignores]
"*/tests/*" = ["S101"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "documentary_watch.settings_test"

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
omit = ["*migrations*", "*tests*"]
plugins = ["django_coverage_plugin"]

[tool.deptry]
#skip_obsolete = true
#skip_missing = true
#skip_transitive = true
#skip_misplaced_dev = true

ignore_obsolete = [
    "crispy-bootstrap5",
    "django-crispy-forms",
    "django-model-utils",
    "whitenoise"

]
ignore_missing = [
    "factory",


]
ignore_transitive = [
    "pytest",

]